backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
apscheduler==3.10.4
alpaca-py==0.20.0
yfinance==0.2.38
pandas==2.2.2
httpx==0.27.0
python-telegram-bot==21.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
